fix(specprefill): preserve input dtype in manual RoPE helpers
manual_rope/manual_rope_with_freqs compute angles in float32 (right for
precision) but returned float32 outputs. Under sparse_prefill the
position-mapped RoPE therefore produced float32 q_pe/k_pe on the target;
on GLM DSA, which passes pe_scores as the fused sdpa mask, that surfaced
as: SpecPrefill sparse prefill failed: [scaled_dot_product_attention]
Mask type must promote to output type bfloat16 -- and the request paid
the draft scoring cost and then fell back to a full dense prefill.
Reproduced end-to-end with GLM-4.7-Flash-4bit drafting for GLM-5.2
oQ4e-mtp on an M3 Ultra (scoring succeeded: 42952 tokens scored in
88.4s, 12872 selected; sparse prefill then failed pre-fix).

Cast the concatenated output back to x.dtype, matching native mx RoPE
semantics. Side effect for all architectures: sparse_prefill no longer
writes silently-upcast float32 roped keys into the KV cache.

+++ omlx/patches/specprefill.py
@@ -712,7 +712,12 @@ def manual_rope(x, positions, dims, base=10000.0, scale=1.0):
     rotated = mx.concatenate(
         [x1 * cos_a - x2 * sin_a, x1 * sin_a + x2 * cos_a], axis=-1
     )
-    return mx.concatenate([rotated, x_pass], axis=-1)
+    # Angles are computed in float32 for precision, but the output must
+    # keep the input dtype like native mx RoPE does: a float32 leak here
+    # becomes float32 roped K in the cache, and on mask-fused MLA targets
+    # (GLM DSA passes pe_scores as the sdpa mask) a float32 mask that
+    # cannot promote to the model's bfloat16 output.
+    return mx.concatenate([rotated, x_pass], axis=-1).astype(x.dtype)
 
 
 def manual_rope_with_freqs(x, positions, dims, freqs, pre_scale=1.0):
@@ -756,7 +761,12 @@ def manual_rope_with_freqs(x, positions, dims, freqs, pre_scale=1.0):
     rotated = mx.concatenate(
         [x1 * cos_a - x2 * sin_a, x1 * sin_a + x2 * cos_a], axis=-1
     )
-    return mx.concatenate([rotated, x_pass], axis=-1)
+    # Angles are computed in float32 for precision, but the output must
+    # keep the input dtype like native mx RoPE does: a float32 leak here
+    # becomes float32 roped K in the cache, and on mask-fused MLA targets
+    # (GLM DSA passes pe_scores as the sdpa mask) a float32 mask that
+    # cannot promote to the model's bfloat16 output.
+    return mx.concatenate([rotated, x_pass], axis=-1).astype(x.dtype)
 
 
 # ---------------------------------------------------------------------------
